refactor tests: move to whitebox and add more use-cases
